More aggressive optimisation of reaction moderation
This makes it so we basically never have to fetch messages being
reacted to UNLESS the reaction emoji is in the ban list, which
CONSIDERABLY reduces the amount of message fetching we need to do for
old messages.

A side effect of this optimisation however is that now we expect emoji
in the configuration file to always be wrapped in brackets if they are
custom Discord emotes. Honestly this is an improvement since what
Discord outputs when you prefix an emoji with \ is an identifier that
is already wrapped in brackets.