Fix moderation of super (burst) reactions
As of discord.py 2.3.2, we are running into an issue where for some
reason burst reactions do not have a user list attached to them, so
we cannot retrieve the users who reacted with a burst reaction.
However, we can work around that using the user id from the raw
reaction event payload even if it may be a bit annoying at times.

This adds some safeguards for situations like this. In this version
of discord.py, it seems like the only way we can tell something MAY be
a burst reaction is by finding out that it has no users attached.
There already is a patch adding proper support for burst reactions
upstream on the horizon (#9309) but no idea if that will fix the user
list issue we are experiencing here.

For the time being, this solution works and allows burst reactions to
be moderated even if the user list may be reported incomplete as a
result.